refactor tests for fixtures in fixtures
